backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy>=2.0
pydantic>=2.7
PyJWT
web3>=6.0.0
eth-account>=0.9.0
siwe>=2.2.0
passlib[bcrypt]
python-multipart
pytest
